+++ CHANGELOG.md
@@ -46,9 +46,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [cli] `--query-timeout` command-line option ([3b658f3]).
 - [cli] `--itc/--iterm2-compress`, `--itjq/--iterm2-jpeg-quality` and `--itnrff/iterm2-no-read-from-file` style-specific command-line options ([#55]).
 - [cli] `-multi` command-line option.
+- [cli] `--swap_win_size` and `--no-swap_win_size` command-line options.
 - [tui] Concurrent/Parallel frame rendering for TUI animations ([#42]).
 - [tui] Included key codes in the help menu.
 - [cli,tui] `no multi`, `query timeout` and `style` config options.
+- [cli,tui] `swap win size` config option.
 - [cli,tui] Attempt to set window title at startup.
 - [lib,cli,tui] Support for the Kitty terminal graphics protocol ([#39]).
 - [lib,cli,tui] Automatic render style selection based on the detected terminal support ([#37]).